update Product API

+++ Bikya.Services/Services/ProductImageService.cs
@@ -1,6 +1,7 @@
 ﻿using Bikya.Data.Models;
 using Bikya.Data.Repositories.Interfaces;
 using Bikya.DTOs.ProductDTO;
+using Microsoft.AspNetCore.Identity;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -14,16 +15,23 @@ namespace Bikya.Services.Services
     {
         private readonly IProductImageRepository _productImageRepository;
         private readonly IProductRepository _productRepository;
- 
+        private readonly UserManager<ApplicationUser> userManager;
 
 
-        public ProductImageService(IProductImageRepository productImageRepository, IProductRepository productRepository)
+        public ProductImageService(IProductImageRepository productImageRepository, IProductRepository productRepository, UserManager<ApplicationUser> userManager)
         {
             _productImageRepository = productImageRepository;
             _productRepository = productRepository;
+            this.userManager = userManager;
         }
 
-       
+        public async Task<bool> ISAdmin(int userId)
+        {
+            var user = await userManager.FindByIdAsync(userId.ToString());
+            var roles = await userManager.GetRolesAsync(user);
+            return roles.Contains("Admin");
+        }
+
         public async Task AddProductImageAsync(ProductImageDTO productImageDTO,string rootPath)
         {
             var product = await _productRepository.GetByIdAsync(productImageDTO.ProductId);
@@ -84,11 +92,30 @@ namespace Bikya.Services.Services
             await _productImageRepository.UpdateAsync(productImage);
         }
 
-        public async Task DeleteProductImageAsync(int id)
+        public async Task DeleteProductImageAsync(int id,int userId,string rootPath)
         {
             var productImage = await _productImageRepository.GetByIdAsync(id);
             if (productImage == null)
                 throw new ArgumentException("Product image not found", nameof(id));
+            var product = await _productRepository.GetByIdAsync(productImage.ProductId);
+            if(product == null)
+                throw new ArgumentException("Product not found");
+
+            var isAdmin = await ISAdmin(userId);
+            if (product.UserId != userId&&!isAdmin)
+                throw new UnauthorizedAccessException("You do not have permission to delete this product image");
+
+            if (product.Status != Data.Enums.ProductStatus.Available)
+                throw new InvalidOperationException("You can not update a product that is in Process");
+
+            //var fileName = productImage.ImageUrl.TrimStart('/');
+            //var filePath = Path.Combine(rootPath,fileName);
+
+            //if (System.IO.File.Exists(filePath))
+            //{
+            //    System.IO.File.Delete(filePath);
+            //}
+
 
             await _productImageRepository.DeleteAsync(productImage);
         }
